refactor: extract domain components
