+++ apps/perion-steam-challenge/src/components/game-count/game-count.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import {GameCount} from './game-count';
+
+describe('GameCount', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<GameCount />);
+    expect(baseElement).toBeTruthy();
+  });
+});
