+++ apps/perion-steam-challenge/src/components/game-list/game-list.module.scss
@@ -1,0 +1,7 @@
+/*
+ * Replace this with your own classes
+ *
+ * e.g.
+ * .container {
+ * }
+*/
