+++ apps/perion-steam-challenge/src/components/game-list/game-list.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import {GameList} from './game-list';
+
+describe('GameList', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<GameList games={[]} />);
+    expect(baseElement).toBeTruthy();
+  });
+});
