+++ apps/perion-steam-challenge/src/components/game-search/game-search.module.scss
@@ -1,0 +1,7 @@
+/*
+ * Replace this with your own classes
+ *
+ * e.g.
+ * .container {
+ * }
+*/
